Svg'es moved in a commen place
